Move AppView and AppViewContainer out of App.js
